requirements.txt
alembic==1.4.2
aniso8601==8.0.0
apispec==4.0.0
apispec-webframeworks==0.5.2
appdirs==1.4.4
arrow==0.14.7
attrs==21.4.0
beautifulsoup4==4.8.2
black==20.8b1
certifi==2019.11.28
cfgv==3.3.1
chardet==3.0.4
click==8.0.3
coverage==6.2
distlib==0.3.4
feedparser==6.0.8
filelock==3.4.2
flake8==3.8.4
Flask==1.1.1
flask-apispec==0.11.0
Flask-Cors==3.0.10
Flask-gzip==0.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
ics==0.7
identify==2.4.4
idna==2.9
iniconfig==1.1.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.24.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
peewee==3.13.1
platformdirs==2.4.1
pluggy==0.13.1
pre-commit==2.9.3
psycopg2-binary==2.8.4
py==1.11.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.1.2
pytest-cov==2.10.1
pytest-datadir==1.3.1
pytest-mock==3.3.1
pytest-split==0.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2019.3
PyYAML==6.0
regex==2022.1.18
requests==2.23.0
requests-mock==1.9.3
robobrowser==0.5.3
sgmllib3k==1.0.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
swagger-spec-validator==2.7.3
TatSu==4.4.0
timestring==1.6.4
toml==0.10.2
tomli==1.2.3
typed-ast==1.5.1
typing_extensions==4.0.1
urllib3==1.25.8
virtualenv==20.13.0
webargs==8.1.0
Werkzeug==1.0.0
